// src/types/user.ts
export interface User {
    user_id: string;    // ID del usuario que mapea con la tabla usuarios
    username: string;   // Nombre de usuario
    name: string;     // Nombre completo
    apellido_p: string; // Apellido paterno
    apellido_m: string; // Apellido materno
    phone: string;   // Teléfono
    email: string;      // Correo electrónico
    avatar_url: string; // URL del avatar
  }
  